Add a test for the \d special command.
